fix: Support fractional tile heights in Tiled importer
Fixed critical bug where tile height property was parsed as integer,
truncating fractional values like 0.1 to 0. Changed TiledTileProperties::height
from int32_t to float and updated parsing to use std::stof().

Changes:
- TiledTileProperties::height: int32_t → float (supports 0.1m floor tiles)
- parse_tile_properties: std::stoi → std::stof for height parsing
- validate_map: Updated height validation to use float comparison
- import_to_world: Removed vertical stacking loop - height is now a mesh property
- log.cpp: Fixed UTF-8 crash by removing _setmode(_O_U8TEXT) calls

Test room now loads successfully with:
- Wall tiles: height 1.0m (Cube.fbx)
- Floor tiles: height 0.1m (FloorTile.fbx)

+++ include/lore/world/tiled_importer.hpp
@@ -14,7 +14,7 @@ struct TileCoord;
 // Parsed Tiled tile properties (custom properties set in Tiled editor)
 struct TiledTileProperties {
     std::string mesh_path;           // Path to 3D mesh (e.g., "meshes/wall_stone.obj")
-    int32_t height = 1;              // Vertical height in tiles (1-8 typical)
+    float height = 1.0f;             // Vertical height in meters (supports fractional values like 0.1)
     std::string collision_type = "box"; // "box", "sphere", "mesh", "none"
     uint32_t material_id = 0;        // PBR material index
     bool walkable = true;            // Can entities walk on this?
